backend/requirements.txt
asgiref==3.7.2 
django==4.2.2 
djangorestframework==3.14.0 
pytz==2023.3 
sqlparse==0.4.4 
typing-extensions==4.6.3
certifi==2023.5.7 
cffi==1.15.1 
charset-normalizer==3.1.0 
cryptography==41.0.1 
defusedxml==0.7.1 
django-templated-mail==1.1.1 
djangorestframework-simplejwt==5.2.2 
djoser==2.2.0 
idna==3.4 
oauthlib==3.2.2 
pycparser==2.21 
pyjwt==2.7.0 
python3-openid==3.2.0 
requests==2.31.0 
requests-oauthlib==1.3.1 
social-auth-app-django==5.2.0 
social-auth-core==4.4.2 
urllib3==2.0.3
Pillow==9.5.0
drf-yasg==1.21.6 
inflection==0.5.1 
packaging==23.1 
pyyaml==6.0 
uritemplate==4.1.1
django-timezone-utils==0.15.0
